+++ test/taExecCommand.spec.js
@@ -9,6 +9,7 @@ describe('taExecCommand', function(){
 			$document = _$document_;
 			contents = angular.element('<div>');
 			contents.append('<p>Test Text</p>');
+			$document.find('body').append(contents);
 			// Fake it till you make it
 			_tempExec = $document[0].execCommand;
 			$document[0].execCommand = function(){
@@ -27,7 +28,11 @@ describe('taExecCommand', function(){
 		}));
 		
 		describe('document execCommand fallback', function(){
-			it('should execute for elements', inject(function(taExecCommand){
+			it('should execute for elements', inject(function(taExecCommand, $window){
+				var sel = $window.rangy.getSelection();
+				var range = $window.rangy.createRangyRange();
+				range.selectNodeContents(contents.find('p')[0]);
+				sel.setSingleRange(range);
 				taExecCommand()('bold', false, null);
 				expect(contents.html()).toBe('<p><b>Test Text</b></p>');
 			}));
@@ -301,14 +306,18 @@ describe('taExecCommand', function(){
 	});
 	
 	describe('handles lists correctly', function(){
-		var taSelectionMock, $document, contents;
+		var taSelectionMock, $document, contents, insertedHtml;
 		beforeEach(function(){
+			insertedHtml = '';
 			taSelectionMock = {
 				element: undefined,
 				getSelectionElement: function (){ return this.element; },
 				getOnlySelectedElements: function(){ return this.element.childNodes; },
 				setSelectionToElementStart: function (){ return; },
-				setSelectionToElementEnd: function (){ return; }
+				setSelectionToElementEnd: function (){ return; },
+				setSelectionAfterElement: function (){ return; },
+				setSelectionBeforeElement: function (){ return; },
+				insertHtml: function(html){ insertedHtml = html; }
 			};
 			
 			module(function($provide){
@@ -328,6 +337,67 @@ describe('taExecCommand', function(){
 					taExecCommand()('insertorderedlist', false, null);
 					expect(element.html()).toBe('<b>Test Text</b>');
 					$document[0].execCommand = _temp;
+				}));
+			});
+			describe('nothing selected', function(){
+				beforeEach(inject(function(taSelection){
+					element = angular.element('<div><p></p></div>');
+					taSelection.element = element.children()[0];
+				}));
+				it('to ol', inject(function(taSelection, taExecCommand){
+					taExecCommand()('insertorderedlist', false, null);
+					expect(element.html()).toBe('<ol><li></li></ol>');
+				}));
+				
+				it('to ul', inject(function(taSelection, taExecCommand){
+					taExecCommand()('insertunorderedlist', false, null);
+					expect(element.html()).toBe('<ul><li></li></ul>');
+				}));
+			});
+			describe('br on line selected', function(){
+				beforeEach(inject(function(taSelection){
+					element = angular.element('<div><p><br></p></div>');
+					taSelection.element = element.children()[0];
+				}));
+				it('to ol', inject(function(taSelection, taExecCommand){
+					taExecCommand()('insertorderedlist', false, null);
+					expect(element.html()).toBe('<ol><li><br></li></ol>');
+				}));
+				
+				it('to ul', inject(function(taSelection, taExecCommand){
+					taExecCommand()('insertunorderedlist', false, null);
+					expect(element.html()).toBe('<ul><li><br></li></ul>');
+				}));
+			});
+			describe('empty ta-bind', function(){
+				beforeEach(inject(function(taSelection){
+					element = angular.element('<div class="ta-bind"></div>');
+					taSelection.element = element[0];
+				}));
+				it('to ol', inject(function(taSelection, taExecCommand){
+					taExecCommand()('insertorderedlist', false, null);
+					expect(insertedHtml).toBe('');
+				}));
+				
+				it('to ul', inject(function(taSelection, taExecCommand){
+					taExecCommand()('insertunorderedlist', false, null);
+					expect(insertedHtml).toBe('');
+				}));
+			});
+			describe('text-only ta-bind', function(){
+				beforeEach(inject(function(taSelection){
+					element = angular.element('<div class="ta-bind">testsomecontent</div>');
+					taSelection.element = element[0];
+					taSelection.getOnlySelectedElements = function(){ return []; };
+				}));
+				it('to ol', inject(function(taSelection, taExecCommand){
+					taExecCommand()('insertorderedlist', false, null);
+					expect(element.html()).toBe('<ol><li>testsomecontent</li></ol>');
+				}));
+				
+				it('to ul', inject(function(taSelection, taExecCommand){
+					taExecCommand()('insertunorderedlist', false, null);
+					expect(element.html()).toBe('<ul><li>testsomecontent</li></ul>');
 				}));
 			});
 			describe('single element selected', function(){